Adjust slippage for zeroEx Solver (gnosis/gp-v2-services#1372)
ZeroEx slippage is quite tight right now with 0.05% 
I think this tightness contributes to the huge costs of failed tx. Hence, I want it to be adaptable.

### Test Plan
*<Explain how you and a reviewer have/intend to test this change>*

+++ crates/shared/src/solver_utils.rs
@@ -30,7 +30,7 @@ impl Slippage {
     }
 
     /// Creates a slippage amount from the specified basis points as number.
-    pub fn number_from_basis_points(basis_points: u16) -> Result<Self> {
+    pub fn number_from_basis_points(basis_points: u32) -> Result<Self> {
         let number_representation = (basis_points as f64) / 10000.;
         Ok(Slippage(number_representation))
     }
